add converage for addcmul
